Return page size as unsigned long
Currently page_size() returns unsigned int value that is after "bitwise
not" is promoted to unsigned long value e.g. in uffd.c
handle_page_fault. Since the value is unsigned promotion is done with 0
MSB that results in lost of MSB pagefault address bits. So make
page_size to return  unsigned long to avoid such situation.

+++ compel/plugins/std/infect.c
@@ -27,7 +27,7 @@ static struct rt_sigframe *sigframe;
  */
 static unsigned __page_size;
 
-unsigned __attribute((weak)) page_size(void)
+unsigned long __attribute((weak)) page_size(void)
 {
 	return __page_size;
 }
